Parser: Hotfix bracket_level
Of course, in unbalanced parenthesis, we can have more closing brackets
than opening brackets, in which case the bracket_level would be
negative.

+++ src/frontend/parser.rs
@@ -403,7 +403,7 @@ pub struct Parser<'flags> {
     current_externs: Vec<String>,
     line_start: usize,
     errors: Vec<ParserError>,
-    bracket_level: usize,
+    bracket_level: i32,
     flags: &'flags Flags,
 }
